Changed build procedure so that MUMPS, metis and interface use 64 bit integers. To do that with MUMPS, one must set C and Fortran compiler flags to set 64 bit integers as default. Changed variable declarations in Julia interface Fortran code from integer(kind=8) to just plain integer, then set compiler default integer type to 64 bit.

+++ src/mumps_cmplx_p.f90
@@ -22,7 +22,7 @@ subroutine init( mumps_par, sym )
 implicit none
 
 TYPE (ZMUMPS_STRUC),intent(inout):: mumps_par
-integer(kind=8),intent(in):: sym  ! =0 unsymmetric, =1 symm. pos def, =2 general symm.
+integer,intent(in):: sym  ! =0 unsymmetric, =1 symm. pos def, =2 general symm.
 
    !mumps_par%COMM = MPI_COMM_WORLD
    mumps_par%SYM = sym  ! =0 unsymmetric, =1 symm. pos def, =2 general symm.
@@ -48,10 +48,10 @@ subroutine convert_to_mumps_format( mumps_par, n, A,jA,iA, ierr )
 implicit none
 
 TYPE (ZMUMPS_STRUC),intent(inout):: mumps_par
-integer(kind=8),intent(in):: n  ! # of rows in A
+integer,intent(in):: n  ! # of rows in A
 complex(kind=8),intent(in):: A(*)
-integer(kind=8),intent(in):: jA(*), iA(n+1)
-integer(kind=8),intent(out):: ierr
+integer,intent(in):: jA(*), iA(n+1)
+integer,intent(out):: ierr
 
 integer nonz, i,j, j1,j2, ind, jcol, istat
 
@@ -120,8 +120,8 @@ subroutine factor_matrix( mumps_par, ooc, ierr )
 implicit none
 
 TYPE (ZMUMPS_STRUC),intent(inout):: mumps_par
-integer(kind=8),intent(in):: ooc ! = 0 in-core factorization, = 1 out-of-core factorization
-integer(kind=8),intent(out):: ierr
+integer,intent(in):: ooc ! = 0 in-core factorization, = 1 out-of-core factorization
+integer,intent(out):: ierr
 
 !mumps_par%icntl(2) = 6  ! output stream for diagnostics
 mumps_par%icntl(4) = 0 ! 1  ! amount of output
@@ -213,7 +213,7 @@ subroutine solve( mumps_par, nrhs, rhs, x, transpose )
 implicit none
 
 TYPE (ZMUMPS_STRUC),intent(inout):: mumps_par
-integer(kind=8),intent(in):: nrhs   ! # of right-hand-sides
+integer,intent(in):: nrhs   ! # of right-hand-sides
 complex(kind=8),intent(in):: rhs(nrhs * mumps_par%N)
 complex(kind=8),intent(out),target:: x(nrhs * mumps_par%N)  ! solution
 logical,intent(in):: transpose  ! if .true. take the transpose
@@ -255,7 +255,7 @@ subroutine solve_sparse_rhs( mumps_par, nrhs, x, transpose )
 implicit none
 
 TYPE (ZMUMPS_STRUC),intent(inout):: mumps_par
-integer(kind=8),intent(in):: nrhs   ! # of right-hand-sides
+integer,intent(in):: nrhs   ! # of right-hand-sides
 complex(kind=8),intent(out),target:: x(nrhs * mumps_par%N)  ! solution
 logical,intent(in):: transpose  ! if .true. take the transpose
